wip, many old features added
This commit adds many things that used to be there and were lost during
code migration to rust. Mainly a good part of the collection, table and
graph generation. Further it adds some support for markdown generation
however this is still not complete so the generated pages are yet not in
a better shape than they were in the previous codebase.