Added preview structures
It became clear that Raw and Preview are inherently different things.
Raw is made to be inputted into the collection while Previews are
smaller versions of the final structures made so that they may be
referenced from other structures and displayed on the output without
copying the entire structure and its references. While the current state
if the conversion basically just copies Raw into Preview their differences
will become apparent with future changes.